Implements 6DOF reprojection into depth space method
Summary:
The commit introduces a new method of calculating an environment depth coordinate to sample. Instead of reprojecting the camera's screen space coordinates, the new method uses world space coordinates. It allows the incorporation of translation into the reprojection process, resulting in more stable coordinates when in motion.
A new API is also surfaced, that allows easier integration of the functionality using macros.
Because it's hard to incorporate z-bias with this approach, API now adds a new way to add z-bias to the calculation formula.

Reviewed By: TudorJude

Differential Revision: D51027973

fbshipit-source-id: 4e1fd636b2e7eda28cb4ed1627c481b39f66662c